requirements.txt
Mastodon.py
click
PyYAML
requests
faker
html2text